+++ hao.c
@@ -509,7 +509,7 @@ static void* hao_thread(void* dummy)
 
   UNREFERENCED(dummy);
 
-  WRMSG(HHC00100, "I", thread_id(), getpriority(PRIO_PROCESS,0), "Hercules Automatic Operator");
+  WRMSG(HHC00100, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), "Hercules Automatic Operator");
 
   /* Wait for panel thread to engage */
   while(!sysblk.panel_init && !sysblk.shutdown && !logger_status() )
@@ -551,7 +551,7 @@ static void* hao_thread(void* dummy)
     }
   }
 
-  WRMSG(HHC00101, "I", thread_id(), getpriority(PRIO_PROCESS,0), "Hercules Automatic Operator");
+  WRMSG(HHC00101, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), "Hercules Automatic Operator");
   return NULL;
 }
