+++ httpserv.c
@@ -808,7 +808,7 @@ struct timeval      timeout;            /* timeout value             */
     SETMODE(USER);
 
     /* Display thread started message on control panel */
-    WRMSG (HHC00100, "I", thread_id(), getpriority(PRIO_PROCESS,0), "HTTP server");
+    WRMSG (HHC00100, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), "HTTP server");
 
     /* make sure root path is built */
     if ( http_root() == NULL )
@@ -928,7 +928,7 @@ http_server_stop:
         hdl_rmsc(http_shutdown, NULL);
 
     /* Display thread started message on control panel */
-    WRMSG(HHC00101, "I", thread_id(), getpriority(PRIO_PROCESS,0), "HTTP server");
+    WRMSG(HHC00101, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), "HTTP server");
 
     sysblk.httptid = 0;
