+++ shared.c
@@ -2475,7 +2475,7 @@ char            threadname[40];
 
     /* This thread will be the shared device thread */
     MSGBUF(threadname, "Shared device(%1d:%04X)", SSID_TO_LCSS(dev->ssid), dev->devnum);
-    WRMSG (HHC00100, "I", thread_id(), getpriority(PRIO_PROCESS,0), threadname);
+    WRMSG (HHC00100, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), threadname);
 
     while (dev->shrdconn)
     {
@@ -2606,7 +2606,7 @@ char            threadname[40];
     dev->shrdtid = 0;
     release_lock (&dev->lock);
 
-    WRMSG(HHC00101, "I", thread_id(), getpriority(PRIO_PROCESS,0), threadname);
+    WRMSG(HHC00101, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), threadname);
 
     return NULL;
 
@@ -2698,7 +2698,7 @@ char                    threadname[40];
     MSGBUF(threadname, "Shared device server %d.%d", SHARED_VERSION, SHARED_RELEASE);
 
     /* Display thread started message on control panel */
-    WRMSG (HHC00100, "I", thread_id(), getpriority(PRIO_PROCESS,0), threadname);
+    WRMSG (HHC00100, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), threadname);
 
     /* Obtain a internet socket */
     lsock = socket (AF_INET, SOCK_STREAM, 0);
@@ -2884,7 +2884,7 @@ char                    threadname[40];
 
     sysblk.shrdtid = 0;
 
-    WRMSG (HHC00101, "I", thread_id(), getpriority(PRIO_PROCESS,0), threadname);
+    WRMSG (HHC00101, "I", (u_long)thread_id(), getpriority(PRIO_PROCESS,0), threadname);
 
     return NULL;
